Move DUASignature from access.py to duasignature.py.
